+++ testcontainers/src/core.rs
@@ -1,6 +1,8 @@
 pub use self::{
     containers::*,
-    image::{ContainerState, ExecCommand, Host, Image, ImageArgs, Port, RunnableImage, WaitFor},
+    image::{
+        ContainerState, ExecCommand, Host, Image, ImageArgs, PortMapping, RunnableImage, WaitFor,
+    },
 };
 
 mod image;
@@ -10,5 +12,6 @@ pub(crate) mod containers;
 pub(crate) mod env;
 pub(crate) mod logs;
 pub(crate) mod macros;
+pub(crate) mod mounts;
 pub(crate) mod network;
 pub(crate) mod ports;
